Finished the first round
Randomly assigned a player to 'win' the lamp.  When that player's turn starts they get the lamp, regardless of whether they choose a wrapped gift or try to steal an unwrapped gift.
Closes #3

Trades are limited so that a player can't immediately steal a gift that was stolen from them.  Also, limited the number of trades that can occur in a single turn to six.
Closes #7

Finished all of the interactions required to complete the first round of Grasmas
Closes #11